// lib/intermediacao/types.ts
export type VendaStatus = 'rascunho' | 'em_processamento' | 'concluida' | 'paga';

export type TipoDocumento = 'cpf' | 'cnpj';

export type CargoBeneficiario = 'Corretor' | 'Gerente' | 'Proprietário' | 'Imobiliária';

export type ParcelaStatus = 'pendente' | 'vencida' | 'paga' | 'cancelada';